bff/pyproject.toml
[project]
name = "tigu-delivery-bff"
version = "0.1.0"
description = "FastAPI bridge for Tigu delivery console"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.109.0,<1.0",
  "uvicorn[standard]>=0.25.0,<1.0",
  "SQLAlchemy>=2.0.23,<3.0",
  "asyncmy>=0.2.9,<1.0",
  "pydantic>=2.5.3,<3.0",
  "pydantic-settings>=2.1.0,<3.0",
  "httpx>=0.25.2,<1.0",
  "redis>=5.0.1,<6.0",
  "python-dotenv>=1.0.0,<2.0",
  "python-jose[cryptography]>=3.3.0,<4.0",
  "passlib[bcrypt]>=1.7.4,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.4,<8.0",
  "pytest-asyncio>=0.23.2,<1.0",
  "coverage>=7.3.2,<8.0",
  "ruff>=0.1.9,<1.0",
  "mypy>=1.8.0,<2.0",
  "types-redis>=4.6.0,<5.0"
]

[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"
